History works now with id instead of code row

+++ src/main/java/ch/luklanis/esscan/PreferencesActivity.java
@@ -359,7 +359,7 @@ public class PreferencesActivity extends PreferenceActivity
         private EditTextPreference mNamePreference;
         private EditTextPreference mIbanPreference;
         private ListPreference mExecutionDayPreference;
-        private int mBankId;
+        private long mBankId;
         private HistoryManager mHistoryManager;
 
         @Override
